ollama_to_JSON/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
jsonschema>=4.0.0
tqdm>=4.64.0
